+++ str/apps/src/LPF.cpp
@@ -6,7 +6,7 @@
  *  @param  current_sensor 現在のセンサ値
  *  @return               フィルタ処理を実行したセンサ値
  */
-std::uint8_t LPF::sensor(std::uint8_t current_sensor)
+std::int32_t LPF::sensor(std::int32_t current_sensor)
 {
   // pre_sensorの初期化
   if(pre_sensor == 0){
@@ -24,7 +24,7 @@ std::uint8_t LPF::sensor(std::uint8_t current_sensor)
  *  @param  current_sensor 現在のセンサ値
  *  @return                フィルタ処理を実行したセンサ値
  */
-std::uint8_t LPF::RCFilter(std::uint8_t current_sensor)
+std::int32_t LPF::RCFilter(std::int32_t current_sensor)
 {
   // RCフィルタにおける係数(通常、0.8 or 0.9)
   constexpr double coefficient = 0.9;
@@ -33,5 +33,5 @@ std::uint8_t LPF::RCFilter(std::uint8_t current_sensor)
   // センサ値の上書き
   pre_sensor = current_sensor;
 
-  return static_cast<std::uint8_t>(filtered);
+  return static_cast<std::int32_t>(filtered);
 }
