+++ str/apps/test/LPFTest.cpp
@@ -11,14 +11,14 @@ namespace etrobocon2018_test {
   TEST(LPF, sensorTest)
   {
     LPF filter;
-    std::uint8_t sensor = 24;  // This is 適当
+    std::int32_t sensor = 24;  // This is 適当
 
     // 前回のセンサ値がない場合は、入力値をそのまま返す
     ASSERT_EQ(sensor, filter.sensor(sensor));
 
     // フィルタ処理
-    std::uint8_t current_sensor = 77; // This is 適当
-    std::uint8_t expected = 0.9 * sensor + 0.1 * current_sensor;
+    std::int32_t current_sensor = 77; // This is 適当
+    std::int32_t expected = 0.9 * sensor + 0.1 * current_sensor;
     ASSERT_EQ(expected, filter.sensor(current_sensor));
   }
 }
